+++ MMM/Sources/Models/Requests/CategoryReqDto.swift
@@ -1,0 +1,8 @@
+//
+//  CategoryReqDto.swift
+//  MMM
+//
+//  Created by geonhyeong on 2023/09/16.
+//
+
+import Foundation
