+++ MMM/Sources/Models/Responses/CategoryResDto.swift
@@ -1,0 +1,8 @@
+//
+//  CategoryResDto.swift
+//  MMM
+//
+//  Created by geonhyeong on 2023/09/16.
+//
+
+import Foundation
