+++ MMM/Sources/ViewController/Category/Views/CategorySectionHeader.swift
@@ -1,50 +1,34 @@
 //
-//  CartegoryCollectionViewCell.swift
+//  CategorySectionHeader.swift
 //  MMM
 //
 //  Created by geonhyeong on 2023/09/16.
 //
 
 import UIKit
-import ReactorKit
 
-final class CartegoryCollectionViewCell: BaseCollectionViewCell, View {
-	typealias Reactor = CategoryCollectionViewCellReactor
+final class CategorySectionHeader: BaseCollectionReusableView {
 	
 	// MARK: - UI Components
 	private lazy var titleLabel = UILabel()
 	private lazy var priceLabel = UILabel()
-	
-	override func layoutSubviews() {
-		super.layoutSubviews()
-	}
-	
-	override func prepareForReuse() {
-		super.prepareForReuse()
-		titleLabel.text = ""
-		priceLabel.text = ""
-	}
-	
-	func bind(reactor: CategoryCollectionViewCellReactor) {
-		bindState(reactor)
-		bindAction(reactor)
-	}
+	private lazy var typeImageView = UIImageView()
 }
 //MARK: - Action
-extension CartegoryCollectionViewCell {
+extension CategorySectionHeader {
 }
 //MARK: - Bind
-extension CartegoryCollectionViewCell {
+extension CategorySectionHeader {
 	// MARK: 데이터 변경 요청 및 버튼 클릭시 요청 로직(View -> Reactor)
-	private func bindAction(_ reactor: CategoryCollectionViewCellReactor) {
+	private func bindAction(_ reactor: CategoryReactor) {
 	}
 	
 	// MARK: 데이터 바인딩 처리 (Reactor -> View)
-	private func bindState(_ reactor: CategoryCollectionViewCellReactor) {
+	private func bindState(_ reactor: CategoryReactor) {
 	}
 }
 //MARK: - Attribute & Hierarchy & Layouts
-extension CartegoryCollectionViewCell {
+extension CategorySectionHeader {
 	override func setAttribute() {
 		super.setAttribute()
 		
@@ -64,7 +48,7 @@ extension CartegoryCollectionViewCell {
 	override func setHierarchy() {
 		super.setHierarchy()
 		
-		contentView.addSubviews(titleLabel, priceLabel)
+		addSubviews(typeImageView, titleLabel, priceLabel)
 	}
 	
 	override func setLayout() {
